requirements.txt
selenium
psycopg2-binary
pandas
matplotlib